requirements.txt
argparse
backend
grpcio==1.24.3
keras
numpy
pandas
tensorflow
scikit-allel
Bio
multiprocessing
